Speed up build process
The JHipster configuration has gradle run npm install and frontend
tests before a junit test is run.

This configuration splits the build tasks into the two concerns:
frontend and backend. With "gradle test" only the backend is compiled
and tested. With "gradle webapp_test" only the frontend is compiled and
tested. The property includeWebapp can be used to build both together
again.